Update to new interaction mode
I didn't like using the encoder, it was too inaccurate and not easy to
use. Instead, I decided to have a "scroll mode" setup where I can press
a button to use the trackball as a scroll wheel itself.

Keeping the encoder library around 'cause I think it's still useful, but
now this project no longer uses it.

TODO: I moved around a bunch of the pins while unwiring the encoder,
will need to update the documentationt to match.