Update documentation and clean main source
This is likely the final commit on this repository since I'm moving on
to other things (and no longer want a callback based implementation for
the PMW3360 sensor).

Just cleaning up the code to remove the encoder references, and adding/
updating documentation for any future readers who might want to
reference the code.